chore(Private Key Reader): create legacy private key reader
